backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
openai==1.12.0
python-dotenv==1.0.0
httpx==0.27.0
python-multipart==0.0.6
reportlab==4.0.4
jira==3.5.1
motor==3.3.1
pymongo==4.6.1
requests==2.31.0
